add: 我的加油站 -- 删除功能 1. 增加我的加油站删除功能 2. 更新增加为修改状态优先 3. 更新查询加油站列表功能

+++ gas-station/src/main/java/com/ruoyi/gas/module/geo/domain/vo/UserStationVO.java
@@ -16,7 +16,7 @@ public class UserStationVO {
 
     private String stationName;
 
-    private Long status;
+    private Integer status;
 
     /** 加油站经营数据最新更新时间 */
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
@@ -38,11 +38,11 @@ public class UserStationVO {
         this.stationName = stationName;
     }
 
-    public Long getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
-    public void setStatus(Long status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
